transcriber/pyproject.toml
[tool.poetry]
name = "transcriber"
version = "0.1.0"
description = "文字起こしAPI"
authors = ["Geshi Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
redis = "^5.0.1"
rq = "^1.15.1"
rq-scheduler = "^0.13.1"
python-magic = "^0.4.27"
python-ulid = "^3.0.0"
transformers = "^4.40.2"
punctuators = "0.0.5"
accelerate = "^0.30.0"
torchaudio = "^2.3.0"
stable-ts = "2.16.0"
datasets = "^2.19.1"
librosa = "^0.10.2"
soundfile = "^0.12.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
